+++ src/components/project-file/ImageSourceViewerProofreader/index.tsx
@@ -5,7 +5,8 @@ import classNames from 'classnames';
 import React, { useEffect, useRef } from 'react';
 import { useIntl } from 'react-intl';
 import { useDispatch, useSelector } from 'react-redux';
-import { DebounceStatus, TranslationUser } from '@/components';
+import { DebounceStatus } from '@/components';
+import { TranslationUser } from '../TranslationUser';
 import { APITranslation } from '@/apis/translation';
 import { FC, File, InputDebounceStatus, Source as ISource } from '@/interfaces';
 import { AppState } from '@/store';
